fix: move check change password to shortcuts
